PYCBC-1531: CollectionQueryIndexManager should raise exception when scope/collection names are set in options
Motivation
==========
The RFC specifies that an InvalidArgumentException should be raised if `collection_name` or `scope_name` are set in the options of CollectionQueryIndexManager operations. Currently, the SDK is ignoring them

Changes
=======
If `scope_name` or `collection_name` are given in the options or in kwargs raise an InvalidArgumentException

Results
=======
All tests in CollectionQueryIndexManagerInvalidOptsTest in FIT pass

Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/197421
Tested-by: Build Bot <build@couchbase.com>

+++ acouchbase/management/queries.py
@@ -340,6 +340,14 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
 
         if not kwargs:
             kwargs = {}
+
+        if kwargs.get('scope_name') or (options and options[0].get('scope_name')):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if kwargs.get('collection_name') or (options and options[0].get('collection_name')):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
+
         kwargs['scope_name'] = self._scope_name
         kwargs['collection_name'] = self._collection_name
         super().create_index(self._bucket_name, index_name, fields, *options, **kwargs)
@@ -362,6 +370,14 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
         """
         if not kwargs:
             kwargs = {}
+
+        if kwargs.get('scope_name') or (options and options[0].get('scope_name')):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if kwargs.get('collection_name') or (options and options[0].get('collection_name')):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
+
         kwargs['scope_name'] = self._scope_name
         kwargs['collection_name'] = self._collection_name
         super().create_primary_index(self._bucket_name, *options, **kwargs)
@@ -389,6 +405,14 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
 
         if not kwargs:
             kwargs = {}
+
+        if kwargs.get('scope_name') or (options and options[0].get('scope_name')):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if kwargs.get('collection_name') or (options and options[0].get('collection_name')):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
+
         kwargs['scope_name'] = self._scope_name
         kwargs['collection_name'] = self._collection_name
         super().drop_index(self._bucket_name, index_name, *options, **kwargs)
@@ -410,6 +434,14 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
         """
         if not kwargs:
             kwargs = {}
+
+        if kwargs.get('scope_name') or (options and options[0].get('scope_name')):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if kwargs.get('collection_name') or (options and options[0].get('collection_name')):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
+
         kwargs['scope_name'] = self._scope_name
         kwargs['collection_name'] = self._collection_name
         super().drop_primary_index(self._bucket_name, *options, **kwargs)
@@ -433,6 +465,14 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
         """
         if not kwargs:
             kwargs = {}
+
+        if kwargs.get('scope_name') or (options and options[0].get('scope_name')):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if kwargs.get('collection_name') or (options and options[0].get('collection_name')):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
+
         kwargs['scope_name'] = self._scope_name
         kwargs['collection_name'] = self._collection_name
         super().get_all_indexes(self._bucket_name, *options, **kwargs)
@@ -453,6 +493,14 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
         """
         if not kwargs:
             kwargs = {}
+
+        if kwargs.get('scope_name') or (options and options[0].get('scope_name')):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if kwargs.get('collection_name') or (options and options[0].get('collection_name')):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
+
         kwargs['scope_name'] = self._scope_name
         kwargs['collection_name'] = self._collection_name
         super().build_deferred_indexes(self._bucket_name, *options, **kwargs)
@@ -485,6 +533,13 @@ class CollectionQueryIndexManager(QueryIndexManagerLogic):
 
         if final_args.get('watch_primary', False):
             index_names.append('#primary')
+
+        if final_args.get('scope_name'):
+            raise InvalidArgumentException('scope_name cannot be set in the options when using the collection-level '
+                                           'query index manager')
+        if final_args.get('collection_name'):
+            raise InvalidArgumentException('collection_name cannot be set in the options when using the '
+                                           'collection-level query index manager')
 
         timeout = final_args.get('timeout', None)
         if not timeout:
